% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detconstruct.R
\name{allequal}
\alias{allequal}
\title{Are All Columns in a Matrix Equal?}
\usage{
allequal(x)
}
\arguments{
\item{x}{matrix with column vectors.}
}
\value{
TRUE if matrix \code{x} has zero or one column, or if all column vectors in matrix \code{x} are equal. FALSE if \code{x} contains at least two different columns.
}
\description{
Check if all column vectors in a matrix are equal.
}
