% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{chi2indeptest}
\alias{chi2indeptest}
\title{Pairwise Mutual Independence Test}
\usage{
chi2indeptest(x, alpha)
}
\arguments{
\item{x}{matrix with \code{n} rows comprised of normalized data between 0 and 1.}

\item{alpha}{significance level.}
}
\value{
Object with test outcomes \code{h[i,j] = h[j,i] = TRUE/FALSE} for \code{1 <= i,j <= n} meaning rejection/acceptance of independence null hypothesis involving rows \code{i} and \code{j} of matrix \code{x}, and p-value or confidence level of acceptance.
}
\description{
Pearson's Chi-square test of pairwise mutual independence.
}
