% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{chi2testlinear}
\alias{chi2testlinear}
\title{Goodness-of-Fit Test for Linear Distributions}
\usage{
chi2testlinear(x, alpha)
}
\arguments{
\item{x}{vector with normalized data between 0 and 1.}

\item{alpha}{significance level.}
}
\value{
Object with test outcome \code{h = TRUE/FALSE} meaning rejection/acceptance of linear null hypothesis, and p-value or confidence level of acceptance.
}
\description{
Composite Pearson's Chi-square test for goodness-of-fit of linear distributions.
}
