% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{chi2testuniform}
\alias{chi2testuniform}
\title{Goodness-of-Fit Test for Uniform Distribution}
\usage{
chi2testuniform(x, alpha)
}
\arguments{
\item{x}{vector with normalized data between 0 and 1.}

\item{alpha}{significance level.}
}
\value{
Object with test outcome \code{h = TRUE/FALSE} meaning rejection/acceptance of uniform null hypothesis, and p-value or confidence level of acceptance.
}
\description{
Pearson's Chi-square test for goodness-of-fit of a uniform distribution.
}
