% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detconstruct.R
\name{de.split}
\alias{de.split}
\title{Split a Distribution Element}
\usage{
de.split(dimens, x, sze, id, mode)
}
\arguments{
\item{dimens}{split dimension.}

\item{x}{data in element given as matrix with \code{d} rows or components and \code{n} columns or samples.}

\item{sze}{vector representing the size of the element.}

\item{id}{element index or identifier.}

\item{mode}{for splitting: \code{> 0} or \code{< 0} for equal-size or -score split, respectively.}
}
\value{
Object containing the properties of the resulting two child distribution elements and the split position within the parent element.
}
\description{
Splits a parent distribution element characterized by \code{x}, \code{sze}, and \code{id} along dimension \code{dimens} based on \code{mode} into two child elements.
}
