% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detleafs.R
\name{det.de}
\alias{det.de}
\title{Extract Distribution Element Characteristics}
\usage{
det.de(det, ind)
}
\arguments{
\item{det}{distribution element tree object resulting from \code{\link{det.construct}}.}

\item{ind}{index of element to extract from \code{det}.}
}
\value{
A list with the element characteristics is returned: \code{p} probability density, \code{theta} element parameters, \code{lb} lower bound, \code{size} of element, \code{div} divisions or splits along dimensions leading to final element.
}
\description{
The function \code{det.de} extracts the distribution element with index \code{ind} from a distribution element tree (DET) generated by the function \code{\link{det.construct}}.
}
