% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detconstruct.R
\name{dimstosplit}
\alias{dimstosplit}
\title{Determine Split Dimension(s)}
\usage{
dimstosplit(x, sze, mode, alphag, alphad)
}
\arguments{
\item{x}{data in element given as matrix with \code{d} rows or components and \code{n} columns or samples.}

\item{sze}{vector representing the element size.}

\item{mode}{element order and split mode as detailed in \code{\link{det.construct}}.}

\item{alphag, alphad}{significance levels for goodness-of-fit and independence tests, respectively, in element refinement or splitting process. \code{alphad} is irrelevant for univariate data \code{x} with \code{d = 1}.}
}
\value{
An object comprised of the split dimension(s) or \code{NA} for no split, and the resulting child element parameters is returned.
}
\description{
Determine the split dimensions of an existing distribution element in the DET refinement-process based on statistical tests.
}
