% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{chi2indeptest}
\alias{chi2indeptest}
\title{Pairwise Mutual Independence Test}
\usage{
chi2indeptest(x, alpha)
}
\arguments{
\item{x}{vector with normalized data between 0 and 1.}

\item{alpha}{significance level.}
}
\value{
Object with test outcome \code{h = TRUE/FALSE} meaning rejection/acceptance of pairwise independence null hypothesis, and p-value or confidence level of acceptance.
}
\description{
Pearson's Chi-square test of pairwise mutual independence.
}
